+++ app/controllers/all/statistics/projects_controller.rb
@@ -1,0 +1,5 @@
+class All::Statistics::ProjectsController < ApplicationController
+  def index
+    @project_statistics = ProjectStatistics.new
+  end
+end