+++ CHANGELOG.md
@@ -28,6 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - New views at `/projects/regional/:region/in-progress` and
   `/projects/regional/:region/completed` to show projects NOT assigned to
   Regional caseworker services, filtered by region
+- User can view statistics page for all project statistics
 
 ## [Release 19][release-19]